Abstract createHistoryEntry
